% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LargeScaleModelingThyroid-package.r
\docType{package}
\name{LargeScaleModelingThyroid}
\alias{LargeScaleModelingThyroid}
\alias{LargeScaleModelingThyroid-package}
\title{LargeScaleModelingThyroid.}
\description{
LargeScaleModelingThyroid.
}

